requirements.txt
fastapi
uvicorn
face_recognition
Pillow
numpy
gradio
loguru
